feat: Improve skipped validation logging (#487)
Right now, there are generic logs in FCS which indicates start/end of a
validation. But if a validation is skipped, there is no log indicating
that. These logs are added in FCS without a context of what happens
inside each validation, which can be confusing. To enhance it further, I
added logs to explain that a validation has been SKIPPED.